backend/requirements.txt
fastapi==0.109.0
asyncpg==0.29.0
sqlmodel==0.0.16
uvicorn==0.27.1
pydantic-settings==2.2.1
alembic==1.13.1
websockets==12.0
aioredis==2.0.1